move the old tables class to a subfolder
