+++ lib/database/sembast/sembast_device_dao.dart
@@ -1,6 +1,6 @@
 import 'package:sembast/sembast.dart';
-import 'package:weforza/database/database_tables.dart';
 import 'package:weforza/database/device_dao.dart';
+import 'package:weforza/database/sembast/database_tables.dart';
 import 'package:weforza/exceptions/exceptions.dart';
 import 'package:weforza/extensions/date_extension.dart';
 import 'package:weforza/model/device/device.dart';
