+++ lib/database/settings_dao.dart
@@ -1,5 +1,5 @@
 import 'package:sembast/sembast.dart';
-import 'package:weforza/database/database_tables.dart';
+import 'package:weforza/database/sembast/database_tables.dart';
 import 'package:weforza/model/settings/settings.dart';
 
 /// This class represents an interface
